05_rag_chatbot_with_memory/requirements.txt
langchain
langchain-openai
python-dotenv
streamlit
faiss-cpu
